server/egglocke/requirements.txt
django
mysqlclient
requests
django-ratelimit
django-simple-captcha
